Stacks.c
This is a basic stack implementation using the functions of stack. Stacks works on principle of FIRST IN LAST OUT. 

push() function is used to add element in the stack where top points to the topmost element in the stack

pop() function is used to delete the top most element from the stack. pop() function returns the deleted value 

peek() function returns the topmost element of the stack

display() function show the elements of the stack

Switch case is used to implement a menu driven program for the user